chore: RetainingLogger composes RetainingRingBuffer; extract ILogger to base to break circular dep
Retire the hand-rolled ring buffer in RetainingLogger (5 private fields +
_toOrderedArray()) in favour of composing RetainingRingBuffer<ILogRecord>
from @fgv/ts-utils/collections. Public API unchanged.

Necessary side-effect: extracted ReporterLogLevel, shouldLog,
stringifyLogValue, ILogger, IDetailLogger, isDetailLogger from
logging/logger.ts into a new base/loggerInterface.ts to break the
circular packlet dependency introduced by logging -> collections
(collections/readOnlyConvertingResultMap.ts already imports ILogger from
logging). Updated collections to import ILogger from base; logging re-
exports the moved symbols for backward compat. api.md change is additive
only (symbols gain a direct top-level export path in addition to the
Logging namespace; no exports removed or renamed).
